refactor(framework-next): dedup image loader, unify swc tier ordering
