+++ packages/framework-next/loaders/next-image-loader-stub.cjs
@@ -41,4 +41,3 @@ const nextImageLoaderStub = function NextImageLoader(content) {
 nextImageLoaderStub.raw = true
 
 module.exports = nextImageLoaderStub
-module.exports.raw = true
